pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-ops-dashboard"
version = "0.1.0"
description = "TUI dashboard for media services, torrents, and host metrics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "textual>=0.86.1",
  "rich>=13.8.1",
  "psutil>=6.1.0",
  "httpx>=0.27.2",
  "python-dotenv>=1.0.1",
  "asciimatics>=1.15.0",
]

[project.scripts]
dashboard = "dashboard.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
